+++ packages/@expo/cli/e2e/__tests__/export-test.ts
@@ -58,13 +58,13 @@ it('runs `npx expo export --help`', async () => {
 
       Options
         <dir>                      Directory of the Expo project. Default: Current working directory
-        --dev                      Configure static files for developing locally using a non-https server
         --output-dir <dir>         The directory to export the static files to. Default: dist
-        --max-workers <number>     Maximum number of tasks to allow the bundler to spawn
-        --dump-assetmap            Dump the asset map for further processing
-        --dump-sourcemap           Dump the source map for debugging the JS bundle
-        -p, --platform <platform>  Options: android, ios, web, all. Default: all
+        --dev                      Configure static files for developing locally using a non-https server
         --no-minify                Prevent minifying source
+        --max-workers <number>     Maximum number of tasks to allow the bundler to spawn
+        --dump-assetmap            Emit an asset map for further processing
+        -p, --platform <platform>  Options: android, ios, web, all. Default: all
+        -s, --source-maps          Emit JavaScript source maps
         -c, --clear                Clear the bundler cache
         -h, --help                 Usage info
     "